reserve/requirements.txt
yadisk = 1.3.3
requests = 2.31.0
progress.bar = 8